rename icustom-widgets
