+++ src/assessments/custom-widgets/test-steps/expected-input.tsx
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import * as React from 'react';
 
-import { CustomWidgetPropertyBag } from 'common/types/property-bag/icustom-widgets';
+import { CustomWidgetPropertyBag } from 'common/types/property-bag/custom-widgets-property-bag';
 import { VisualizationType } from 'common/types/visualization-type';
 import { link } from 'content/link';
 import { productName } from 'content/strings/application';
