+++ src/common/types/property-bag/custom-widgets-property-bag.ts
